Added All Code from Game.
The code for this game is very old. I reimported the game into the newest version of Unity, and added some quality of life features like being able to exit the game. I am not proud of the code (Too much in the update function), but it is still a project I worked on, and I am proud of the progress I have made coding since then.